requirements.txt
python-docx
latex2mathml
mathml2omml